Add user trade history, Version 1.2.0

+++ code/model/material_category.class.php
@@ -11,20 +11,30 @@ class material_categorymodel {
         $this->db = $base->db;
     }
 
+    // JUST FOR FIXED THE category INFO
+    function get_list() {
+        $query = $this->db->query("SELECT * FROM `material_category`");
+        $category_list = array();
+        while ($category = $this->db->fetch_array($query)) {
+            $category_list[] = $category;
+        }
+        return $category_list;
+    }
+
     function get_by_mid($mid, $limit='') {
-        $sql = "SELECT `cid` FROM `material_category` WHERE material_id=$mid";
+        $sql = "SELECT * FROM `material_category` WHERE material_id=$mid";
         !empty($limit) && $sql .=  " LIMIT 0,$limit ";
 
         $query = $this->db->query($sql);
-        $cid_list = array();
-        while ($cid = $this->db->fetch_array($query)) {
-            $cid_list[] = $cid['cid'];
+        $category_list = array();
+        while ($category = $this->db->fetch_array($query)) {
+            $category_list[] = $category;
         }
-        return $cid_list;
+        return $category_list;
     }
 
-    function get_by_cid($cid, $start=0, $limit='') {
-        $sql = "SELECT `material_id` FROM `material_category` WHERE cid='$cid'";
+    function get_by_cid($cid, $type="major_id", $start=0, $limit='') {
+        $sql = "SELECT `material_id` FROM `material_category` WHERE $type='$cid'";
         !empty($limit) && $sql .=  " LIMIT $start,$limit ";
 
         $query = $this->db->query($sql);
@@ -35,8 +45,8 @@ class material_categorymodel {
         return $mid_list;
     }
 
-    function get_full_by_cid($cid, $start=0, $limit=10) {
-        $sql = "SELECT * FROM `material` WHERE `id` IN (SELECT `material_id` FROM `material_category` WHERE cid='$cid' LIMIT $start,$limit)";
+    function get_full_by_cid($cid, $type="major_id", $start=0, $limit=10) {
+        $sql = "SELECT * FROM `material` WHERE `id` IN (SELECT `material_id` FROM `material_category` WHERE $type='$cid' LIMIT $start,$limit)";
 
         $query = $this->db->query($sql);
         $material_list = array();
@@ -47,27 +57,57 @@ class material_categorymodel {
         return $material_list; 
     }
 
-    function get_cid_material_num($cid) {
-        return $this->db->result_first("SELECT COUNT(*) FROM material_category WHERE cid='$cid'");
+    function get_cid_material_num($cid, $type="major_id") {
+        return $this->db->result_first("SELECT COUNT(*) FROM material_category WHERE $type='$cid'");
     }
 
-    function multi_add($mid, $cid_list) {
-        //print_r($cid_list);
+    function multi_add($mid, $cid_list, $delete_old=1) {
         if (empty($cid_list)) {
             return false;
         }
-        $this->db->query("DELETE FROM material_category WHERE mid=$mid");
-        $insertsql = "INSERT INTO material_category(`material_id`,`cid`) VALUES ";
+        if ($delete_old) {
+            $this->db->query("DELETE FROM material_category WHERE mid=$mid");
+        }
+
+        $insertsql = "INSERT INTO material_category(`material_id`,`region_id`,`school_id`,`dept_id`,`major_id`) VALUES ";
         foreach ($cid_list as $cid) {
-            $insertsql .= "('$mid','$cid'),";
+            $insertsql .= "('$mid','" . $cid['region_id'] . "','" . $cid['school_id'] . "','" . $cid['dept_id'] . "','" . $cid['major_id'] . "'),";
         }
         $this->db->query(substr($insertsql, 0, -1));
         return true;
     }
 
-    function remove_by_cid($cid_list) {
-        $namestr = "'" . implode("','", $cid_list) . "'";
-        $this->db->query("DELETE FROM material_category WHERE `cid` IN ($namestr)");
+    function remove_by_cid($cid_list, $type="major_id") {
+        $cidstr = "'" . implode("','", $cid_list) . "'";
+        $this->db->query("DELETE FROM material_category WHERE `$type` IN ($cidstr)");
+
+        return $this->db->affected_rows();
+    }
+
+    function remove_by_mid_majorid($mid, $major_id) {
+        $this->db->query("DELETE FROM material_category WHERE `material_id`='$mid' AND `major_id`='$major_id'");
+        return $this->db->affected_rows();
+    }
+
+    function get_majorid_info($major_id_list) {
+        $major_info_list = array();
+
+        $sql = "SELECT * FROM `material_category_info` WHERE major_id IN (";
+        foreach ($major_id_list as $major_id) {
+            $sql .= "'$major_id',";
+        }
+        $sql = substr($sql, 0, -1) . ")";
+
+        $query = $this->db->query($sql);
+        while ($major_info = $this->db->fetch_array($query)) {
+            $major_info_list[$major_info['major_id']] = $major_info;
+        }
+        return $major_info_list;
+    }
+
+    function update($id, $material_id, $region_id, $school_id, $dept_id, $major_id) {
+        $this->db->query("UPDATE `material_category` SET `material_id`='$material_id',`region_id`='$region_id',`school_id`='$school_id',`dept_id`='$dept_id',`major_id`='$major_id' WHERE `id`='$id'");
+        return $this->db->affected_rows();
     }
 }
